pyproject.toml
[project]
name = "tcc-mba-usp-esalq-dsa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasus-fetcher>=0.3.3",
    "geopandas>=1.0.1",
    "pyarrow>=18.0.0",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
